test_requirements.txt
coverage
coveralls
django-nose>=1.2
mock>=1.0.1
nose>=1.3.0
